Building a channel model for MIMO system
Adaptive antennas are used in mobile radio systems for improving coverage and increasing capacity. For realistic system design and simulation, channel models are required in order to assess the performance of such systems. We propose the so-called ‘geometry-based stochastic channel model’ (GSCM), which is easy to implement and has a low complexity even when the directional dimension of the channel is taken into account. In this model, we prescribe the probability density function (PDF) of the location of the scatterers. From this, the angularly resolved impulse response can be computed by a simple ray tracing